added a check

+++ raspi_cam.js
@@ -30,17 +30,21 @@ var upload_to_s3 = function (file_to_put) {
     var target_image_path = '/images/haworth/' + file_to_put;
 
     var s3_upload = s3client.putFile(file_to_put, target_image_path, function (err, s3upres) {
-        console.log('inside s3_upload callback...');
         if (err) {
             console.error("s3 put error...");
             console.error(err);
         }
-        if (s3upres.statusCode == 200) {
-            console.log('finished uploading ' + file_to_put + ' to s3!');
-            fs.unlink(file_to_put);
+        if (s3upres) {
+            if (s3upres.statusCode == 200) {
+                console.log('finished uploading ' + file_to_put + ' to s3!');
+                fs.unlink(file_to_put);
+            }
+            else {
+                console.log(s3upres.statusCode);
+            }
         }
         else {
-            console.log(s3upres.statusCode);
+            console.log('s3upres is empty');
         }
     });
 }
@@ -59,7 +63,7 @@ function takeShot(params) {
         fswebcam = spawn('fswebcam', [
             '--device', '/dev/video0', 
             '--no-banner', 
-            '--nodrop', '1',
+            '--nodrop',
             '--resolution', '1280x720', 
             '--jpeg', '100', 
             '--save', filename
